move WandbConfig
